# Conflicts:
#	docker-compose.yaml
#	src/main/java/mipt/app/secondmemory/controller/file/FilesController.java
#	src/main/java/mipt/app/secondmemory/controller/file/FilesControllerImpl.java
#	src/main/java/mipt/app/secondmemory/service/FilesService.java

+++ src/main/java/mipt/app/secondmemory/controller/file/FilesControllerImpl.java
@@ -1,5 +1,9 @@
 package mipt.app.secondmemory.controller.file;
 
+import static mipt.app.secondmemory.entity.RoleType.OWNER;
+import static mipt.app.secondmemory.entity.RoleType.READER;
+import static mipt.app.secondmemory.entity.RoleType.WRITER;
+
 import io.minio.Result;
 import io.minio.errors.ErrorResponseException;
 import io.minio.errors.InsufficientDataException;
@@ -8,11 +12,10 @@ import io.minio.errors.InvalidResponseException;
 import io.minio.errors.ServerException;
 import io.minio.errors.XmlParserException;
 import io.minio.messages.Item;
-
+import jakarta.servlet.http.Part;
 import java.io.IOException;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
-import jakarta.servlet.http.Part;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -20,6 +23,8 @@ import mipt.app.secondmemory.dto.directory.DirectoryInfoRequest;
 import mipt.app.secondmemory.dto.directory.RootDirectoriesRequest;
 import mipt.app.secondmemory.dto.file.FileInfoRequest;
 import mipt.app.secondmemory.dto.file.FileInfoResponse;
+import mipt.app.secondmemory.entity.Role;
+import mipt.app.secondmemory.entity.Session;
 import mipt.app.secondmemory.exception.directory.BucketNotFoundException;
 import mipt.app.secondmemory.exception.directory.NoSuchBucketException;
 import mipt.app.secondmemory.exception.directory.NoSuchDirectoryException;
@@ -27,8 +32,13 @@ import mipt.app.secondmemory.exception.file.DatabaseException;
 import mipt.app.secondmemory.exception.file.FileAlreadyExistsException;
 import mipt.app.secondmemory.exception.file.FileMemoryLimitExceededException;
 import mipt.app.secondmemory.exception.file.FileNotFoundException;
+import mipt.app.secondmemory.exception.role.NoRoleFoundException;
+import mipt.app.secondmemory.exception.session.SessionNotFoundException;
+import mipt.app.secondmemory.repository.RolesRepository;
+import mipt.app.secondmemory.repository.SessionsRepository;
 import mipt.app.secondmemory.service.FilesService;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.authorization.AuthorizationDeniedException;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -38,9 +48,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1")
 public class FilesControllerImpl implements FilesController {
   private final FilesService filesService;
+  private final RolesRepository rolesRepository;
+  private final SessionsRepository sessionsRepository;
 
   @Override
-  public ResponseEntity<FileInfoResponse> uploadFile(Long bucketId, Part file)
+  public ResponseEntity<FileInfoResponse> uploadFile(Long bucketId, Part file, String cookieValue)
       throws ServerException,
           InsufficientDataException,
           ErrorResponseException,
@@ -51,14 +63,15 @@ public class FilesControllerImpl implements FilesController {
           NoSuchBucketException,
           InvalidResponseException,
           XmlParserException,
-          InternalException,
-          BucketNotFoundException {
-
-    return ResponseEntity.ok(filesService.uploadFile(bucketId, file));
+          InternalException {
+    Session session =
+        sessionsRepository.findByCookie(cookieValue).orElseThrow(SessionNotFoundException::new);
+    return ResponseEntity.ok(filesService.uploadFile(bucketId, file, session.getUser()));
   }
 
   @Override
-  public ResponseEntity<FileInfoResponse> renameFile(Long fileId, String newFileName)
+  public ResponseEntity<FileInfoResponse> renameFile(
+      Long fileId, String newFileName, String cookieValue)
       throws ServerException,
           InsufficientDataException,
           FileNotFoundException,
@@ -71,11 +84,25 @@ public class FilesControllerImpl implements FilesController {
           InternalException,
           NoSuchDirectoryException,
           BucketNotFoundException {
+    //    return ResponseEntity.ok(filesService.renameFile(fileId, newFileName));
+    //          InternalException,
+    //          SessionNotFoundException {
+    Session session =
+        sessionsRepository.findByCookie(cookieValue).orElseThrow(SessionNotFoundException::new);
+    Role userRole =
+        rolesRepository
+            .findByUserIdAndFileId(session.getUser().getId(), fileId)
+            .orElseThrow(NoRoleFoundException::new);
+    if (userRole.getType() != WRITER && userRole.getType() != OWNER) {
+      throw new AuthorizationDeniedException("You don't have permission to rename file");
+    }
     return ResponseEntity.ok(filesService.renameFile(fileId, newFileName));
+    //    filesService.renameFile(bucketName, oldKey, newKey);
+    //    return ResponseEntity.ok().build();
   }
 
   @Override
-  public ResponseEntity<FileInfoResponse> deleteFile(Long fileId)
+  public ResponseEntity<FileInfoResponse> deleteFile(Long fileId, String cookieValue)
       throws ServerException,
           InsufficientDataException,
           FileNotFoundException,
@@ -89,11 +116,25 @@ public class FilesControllerImpl implements FilesController {
           NoSuchDirectoryException,
           BucketNotFoundException {
 
+    //    return ResponseEntity.ok(filesService.deleteFile(fileId));
+    //          InternalException,
+    //          SessionNotFoundException {
+    Session session =
+        sessionsRepository.findByCookie(cookieValue).orElseThrow(SessionNotFoundException::new);
+    Role userRole =
+        rolesRepository
+            .findByUserIdAndFileId(session.getUser().getId(), fileId)
+            .orElseThrow(NoRoleFoundException::new);
+    if (userRole.getType() != OWNER) {
+      throw new AuthorizationDeniedException("You don't have permission to delete file");
+    }
+    //    filesService.deleteFile(bucketName, key);
     return ResponseEntity.ok(filesService.deleteFile(fileId));
+    //    return ResponseEntity.ok().build();
   }
 
   @Override
-  public ResponseEntity<FileInfoResponse> moveFile(Long fileId, Long folderId)
+  public ResponseEntity<FileInfoResponse> moveFile(Long fileId, Long folderId, String cookieValue)
       throws ServerException,
           InsufficientDataException,
           FileNotFoundException,
@@ -109,6 +150,18 @@ public class FilesControllerImpl implements FilesController {
           NoSuchDirectoryException,
           BucketNotFoundException {
 
+    //    return ResponseEntity.ok(filesService.moveFile(fileId, folderId));
+    //          NoSuchBucketException,
+    //          SessionNotFoundException {
+    Session session =
+        sessionsRepository.findByCookie(cookieValue).orElseThrow(SessionNotFoundException::new);
+    Role userRole =
+        rolesRepository
+            .findByUserIdAndFileId(session.getUser().getId(), fileId)
+            .orElseThrow(NoRoleFoundException::new);
+    if (userRole.getType() != OWNER) {
+      throw new AuthorizationDeniedException("You don't have permission to move file");
+    }
     return ResponseEntity.ok(filesService.moveFile(fileId, folderId));
   }
 
@@ -134,20 +187,34 @@ public class FilesControllerImpl implements FilesController {
   }
 
   @Override
-  public ResponseEntity<FileInfoResponse> getFileInfo(long fileId, FileInfoRequest fileInfoRequest)
-      throws FileNotFoundException, DatabaseException {
+  public ResponseEntity<FileInfoResponse> getFileInfo(
+      long fileId, FileInfoRequest fileInfoRequest, String cookieValue)
+      throws FileNotFoundException, DatabaseException, SessionNotFoundException {
+    Session session =
+        sessionsRepository.findByCookie(cookieValue).orElseThrow(SessionNotFoundException::new);
+    Role userRole =
+        rolesRepository
+            .findByUserIdAndFileId(session.getUser().getId(), fileId)
+            .orElseThrow(NoRoleFoundException::new);
+    if (userRole.getType() != READER
+        && userRole.getType() != WRITER
+        && userRole.getType() != OWNER) {
+      throw new AuthorizationDeniedException("You don't have permission to read file information");
+    }
     return ResponseEntity.ok(filesService.getFileInfo(fileId, fileInfoRequest));
   }
 
   @Override
   public ResponseEntity<Iterable<Result<Item>>> getFilesInDirectory(
-      DirectoryInfoRequest directoryInfoRequest) throws NoSuchDirectoryException {
+      DirectoryInfoRequest directoryInfoRequest, String cookieValue)
+      throws NoSuchDirectoryException {
     return ResponseEntity.ok(filesService.getFilesInDirectory(directoryInfoRequest));
   }
 
   @Override
   public ResponseEntity<Iterable<Result<Item>>> getRootDirectories(
-      RootDirectoriesRequest rootDirectoriesRequest) throws NoSuchBucketException {
+      RootDirectoriesRequest rootDirectoriesRequest, String cookieValue)
+      throws NoSuchBucketException {
     return ResponseEntity.ok(filesService.getRootDirectories(rootDirectoriesRequest));
   }
 }
